Update example fn
